fix assertion in profile data writer

+++ llvm/lib/ProfileData/InstrProfWriter.cpp
@@ -75,7 +75,7 @@ public:
       M += sizeof(uint64_t); // The size of the Counts vector
       M += ProfRecord.Counts.size() * sizeof(uint64_t);
       M += sizeof(uint64_t); // The size of the Bitmap vector
-      M += ProfRecord.BitmapBytes.size() * sizeof(uint64_t);
+      M += ProfRecord.BitmapBytes.size() * sizeof(uint8_t);
 
       // Value data
       M += ValueProfData::getSize(ProfileData.second);
@@ -106,8 +106,8 @@ public:
         LE.write<uint64_t>(I);
 
       LE.write<uint64_t>(ProfRecord.BitmapBytes.size());
-      for (uint64_t I : ProfRecord.BitmapBytes)
-        LE.write<uint64_t>(I);
+      for (uint8_t I : ProfRecord.BitmapBytes)
+        LE.write<uint8_t>(I);
 
       // Write value data
       std::unique_ptr<ValueProfData> VDataPtr =
